+++ src/ui/composition.js
@@ -40,18 +40,7 @@ import {
   setDetectionStyle,
   setDetectionTuning,
 } from '../data/detection.js';
-import trafficLayer from '../data/traffic.js';
-import flightsLayer from '../data/flights.js';
-import militaryFlightsLayer from '../data/militaryFlights.js';
 import { isTr3b, toggleTr3b } from '../data/tr3bRegistry.js';
-import satellitesLayer from '../data/satellites.js';
-import cctvLayer from '../data/cctv.js';
-import radioLayer from '../data/radio.js';
-import bikeshareLayer from '../data/bikeshare.js';
-import aisLiveVesselsLayer from '../data/aisLiveVessels.js';
-import militaryAwarenessLayer from '../data/militaryAwareness.js';
-import militaryInstallationsLayer from '../data/militaryInstallations.js';
-import rocketLaunchesLayer from '../data/rocketLaunches.js';
 import {
   cachedGroundFloor,
   cachedMeshFloor,
@@ -114,19 +103,8 @@ export class StyleManager extends ApplicationShell {
         resumeDetection,
         setDetectionStyle,
         setDetectionTuning,
-        trafficLayer,
-        flightsLayer,
-        militaryFlightsLayer,
         isTr3b,
         toggleTr3b,
-        satellitesLayer,
-        cctvLayer,
-        radioLayer,
-        bikeshareLayer,
-        aisLiveVesselsLayer,
-        militaryAwarenessLayer,
-        militaryInstallationsLayer,
-        rocketLaunchesLayer,
         cachedGroundFloor,
         cachedMeshFloor,
         GROUND_FLOOR_LIFT_M,